+++ FarmManager.Model/Model/Employee.cs
@@ -8,7 +8,7 @@ public class Employee : BaseEntity, IContactable
     public string LastName { get; set; } = string.Empty;
     public string? IdNumber { get; set; }
     public bool IsRentable { get; set; } = false;
-    public double? BaseRent { get; set; }
+    public decimal? BaseRent { get; set; }
     public string? PhoneNumber { get; set; }
     public string? Email { get; set; }
 }
