+++ FarmManager.Model/Model/SellHarvestQuantity.cs
@@ -7,5 +7,5 @@ public class SellHarvestQuantity : BaseEntity
     public int SellId { get; set; }
     public Sell Sell { get; set; } = null!;
     public int HarvestId { get; set; }
-    public double Quantity { get; set; }
+    public decimal Quantity { get; set; }
 }
