+++ FarmManager/Models/Workdays/WorkdaysHourly/WorkdayHourlyAddAllModel.cs
@@ -6,6 +6,6 @@ public class WorkdayHourlyAddAllModel
 {
     public ICollection<WorkdayHourly> WorkdaysHourly { get; set; } = [];
     public ICollection<Employee> Employees { get; set; } = [];
-    public double Price { get; set; }
-    public double Hours { get; set; }
+    public decimal Price { get; set; }
+    public decimal Hours { get; set; }
 }
